requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
pymupdf
python-docx
pytesseract
pillow
